Bug fixes, round 1
Most of the way there. Still need to fix an annoying resampling issue
in StreamingBootstrapProposal as well as figure out why the last 3 runs
don't result in any new matchings

+++ src/src/smc/CompactPopulation.cpp
@@ -24,7 +24,7 @@
 
 using namespace sgm;
 
-int smc::CompactPopulation::num_particles() const {
+size_t smc::CompactPopulation::num_particles() const {
     return m_num_particles;
 }
